0.9.2 - A solution to the problem
see the todo but i may have fixed the issue with the page not updating my code after initial reload. This may not be so much of an issue in deployment but in development It's a nightmare. I don't know if this is correct but it works and i can finally move on